refactor(Player): turned double to float

+++ src/Player.h
@@ -18,7 +18,7 @@ public:
     float height;
     int turnDirection;     /** -1 for left, +1 for right */
     int walkDirection;     /** -1 for back, +1 for front */
-    double rotationAngle;
+    float rotationAngle;
     float walkSpeed;
     float turnSpeed;
 };
